Refactor grabbed_device into a subpackage
- Move grabbed_device_*.py into grabbed_device/ package
- Replace module-level aliases with __init__.py exports
- Update all imports
