+++ tests/keymasqd/test_device_manager_runtime_recovery.py
@@ -16,11 +16,11 @@ from keymasq.keymasqd.runtime import actions as adm
 from keymasq.keymasqd.runtime import adapters as runtime_adapters
 from keymasq.keymasqd.runtime import combos as cdm
 from keymasq.keymasqd.runtime import grab_lifecycle as ldm
-from keymasq.keymasqd.runtime import grabbed_device as gdm
-from keymasq.keymasqd.runtime import grabbed_device_actions as gda
-from keymasq.keymasqd.runtime import grabbed_device_events as gde
-from keymasq.keymasqd.runtime import grabbed_device_outputs as gdo
 from keymasq.keymasqd.runtime import topology as tdm
+from keymasq.keymasqd.runtime.grabbed_device import actions as gda
+from keymasq.keymasqd.runtime.grabbed_device import device as gdm
+from keymasq.keymasqd.runtime.grabbed_device import events as gde
+from keymasq.keymasqd.runtime.grabbed_device import outputs as gdo
 from keymasq.keymasqd.superkey_state import SuperkeyActionData, SuperkeyConfig
 from tests.keymasqd.device_manager_support import (
     FakeUInput,
