+++ tests/keymasqd/test_grabbed_device_runtime_helpers.py
@@ -19,13 +19,13 @@ from keymasq.keymasqd import device_manager as dm
 from keymasq.keymasqd.device_manager import DeviceManager
 from keymasq.keymasqd.runtime import adapters as runtime_adapters
 from keymasq.keymasqd.runtime import grab_lifecycle as ldm
-from keymasq.keymasqd.runtime import grabbed_device as gdm
-from keymasq.keymasqd.runtime import grabbed_device_actions as gda
-from keymasq.keymasqd.runtime import grabbed_device_events as gde
-from keymasq.keymasqd.runtime import grabbed_device_grab as gdg
-from keymasq.keymasqd.runtime import grabbed_device_outputs as gdo
-from keymasq.keymasqd.runtime import grabbed_device_repeat as gdr
-from keymasq.keymasqd.runtime import grabbed_device_types as gdt
+from keymasq.keymasqd.runtime.grabbed_device import actions as gda
+from keymasq.keymasqd.runtime.grabbed_device import device as gdm
+from keymasq.keymasqd.runtime.grabbed_device import events as gde
+from keymasq.keymasqd.runtime.grabbed_device import grab as gdg
+from keymasq.keymasqd.runtime.grabbed_device import outputs as gdo
+from keymasq.keymasqd.runtime.grabbed_device import repeat as gdr
+from keymasq.keymasqd.runtime.grabbed_device import types as gdt
 from keymasq.keymasqd.superkey_state import SuperkeyActionData, SuperkeyConfig
 from tests.keymasqd.device_manager_support import (
     FakeUInput,
